fix-single-window-and-temp-folder
removed GUI library and put content in to main APP module
fixed issue with saving temporary csv file in temp directory. previously CNN could not be tested due to non existing path
